rename image mapping file and move to header image component folder from app/
