+++ src/components/HeaderImage/imageMapping.js
